Fix no workflow error on watchlist check produced risk signals (#7016)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

We are now producing risk signals from our continuous monitoring
WatchlistCheckTask. But if you click on those risk signals in the UI, we
will throw an error because currently we are querying for an OBC via the
WF associated with the risk signal but for these WatchlistCheck produced
risk signals, there is no Workflow.
Changes this route to instead query for OBC via a different way. This
route will now just do the same query the WatchlistCheckTask does which
is to take the first OBC for any completed Workflow that is
enhanced_aml=Yes, else take the OBC from the first workflow. This logic
is a bit funky but atm we unfortunately have this weird setup where
continuous monitoring is configured via a playbook but a single vault
can run through workflows on various playbooks.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->