Always filter watchlist risk signals based on enhanced_aml (#7017)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Right now, we filter the list of `hits` we get from an Incode watchlist
call to only those that have a type that corresponds to a type that is
configured in `obc.enhanced_aml`. But a particular `hit` can have
multiple types. So for eg 1 hit could have both types `pep` and
`adverse_media` but a Tenant is only configured to want `pep` hits.
Because of this, we need an additional level of filtering logic that
ensures we only produce risk signals for the relevant types that a
Tenant cares about.
We've been doing this filter in a workflow-specific util, but this moves
it into the Incode watchlist code itself so that the WatchlistCheckTask
is guaranteed to get the same filtering logic too.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->