Add ob-config to verify machine context (#7021)
## Describe your changes

Added ob-config to the machine context.

That being said, this is an optimistic way of doing it; and I believe we
need to update it. The prop `obConfig` is optional in
`GetSdkArgsResponse`. That is it's possible that it will `undefined` in
the payload for `sdkArgsReceived` event. We don't want to proceed email
identification if it's undefined. The ideal approach here is to go to
`<Init/>` and check if the obConfig is already received - if not, make a
request for the args; and if yes send an event to the machine to proceed
to email identification. On the machine, we should have two events such
as `sdkArgsReceived` and `proceed` (we can use a different name) where
`sdkArgsReceived` will only populate the machine context with obConfig,
but not transition and `proceed` will transition with conditional to
check whether obConfig has been populated. We can't currently do it
because the `<Init/>` starts fetching args on-mount because of how it's
set up.
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->