Refactor how we write ContactInfo to the vault (#6942) (#7002)
## Describe your changes
Re-applying this commit that was reverted in
5d3d625391b6d7cd93123d2621ca6c9f941493a9 because of
https://onefootprint.slack.com/archives/C03DQ288QQM/p1700598416008549

This fixes the bug by stopping to mutate data at all.

This was truly just a logic bug in a very niche condition where (1) a
piece of ContactInfo existed in the vault and we're updating it and (2)
the existing ContactInfo was _not_ verified. The really sad part is the
PR on top of this undoes this logic - it was an unncessary in-between :(

As a vaulting company, we should generally be extremely wary of
codepaths that remove data from a request to vault something - one of
the worst things we could do is silently fail to write data that a
tenant gave to us in confidence that we would store it 😞 luckily, when
this bug was deployed, a later assertion saved us from silently
erroring, somewhat by a stroke of luck

To review, take a look at the most recent commit in the PR - that's the
commit that fixes the bug. the bottom two commits are just cherry
picking some logic that was reverted

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->