[FP-6661] Prefill data from portable view into tenant when making new workflow (#6943)
## Describe your changes
The goal is that this makes a change to the write path that is
imperceptible to the read path - we'll start prefilling data when
workflows are created (for the very few one-clicks that happen today),
which would generate a view of the world that is no different from
today.

When this code that is writing the data into the tenant's view is
deployed, we will backfill all data that is being read from another
tenant into the reading tenant, and then we will deprecate the ability
to read data added by another tenant.

## Caveats
The risk of this change is generally low (famous last words) because
there are so few one clicks.

Still many caveats, but among many:
- For simplification right now, we only copy in data on the first
workflow made at a tenant. i think we can rm this soon and copy in every
time
- Still need to adjust the way we compute the portable view of data (the
"n+1"th view as we've been calling it). It's actually buggy as-is today,
but it isn't really used in many places (just the my1fp API).
- We're now copying the phone number from every tenant onto which the
user onboards. This is a little funky since the phone number VaultData
row has some metadata stored in the ContactInfo table, and there's a
unique ContactInfo row for each of the phone number VaultData rows... so
if the verification status of a piece of contact info changes, right now
it will only affect one tenant (not all). in the future, we should
migrate to a representation where contact info exists at the user level
rather than tenant level
- We fingerprint the portable data with a network request to the
enclave, so this happens outside of a locked txn. So by the time we
write the prefill data, it's possible it's stale. Not the end of the
world - the tenant just may get a view of the user that is milliseconds
stale if there were a concurrent update

## Test Plan
Existing portability integration tests
Before landing this, i might test it against a future change that would
have tenants' views of the VW only read data that exists at the tenant