Change CIP risk_signal query (#7043)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Right now in the CIP endpoint we query for RiskSignal's using an old
legacy query that queries for risk signals by OBD, using the
risk_signal<>OBD junction table. I think this is a bit sketch- for eg:
if we manually marked a user as pass in the dashboard, then that would
be the latest OBD the CIP endpoint queries. However, it would retrieve
no RiskSignal's because that OBD would not be associated with any risk
signals because it was just a manual dashboard triggered OBD (ie, we'd
write an empty vec![] for the junction table in that case).
This PR changes the CIP endpoint to just our canonical
`latest_by_risk_signal_group_kinds` query which is what powers dashboard
and such.


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->