Support dry runs in custom migrations (#7047)
## Describe your changes
I was initially going to run the migration via a custom migration, but a
private protected API with various options is much much easier to
control. So, i'm going to do that instead.

However, this is still a useful functionality on custom migrations :)

I'm thinking in the future, we might want to deprecate custom migrations
and just encourage that they are run manually through private protected
APIs

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->