Change CIP vres query (#7044)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Right now the CIP endpoint queries for vres using the
`get_latest_requests_and_successful_results_for_scoped_user` query. But
this query will return a Vreq that has no Vres, if that vreq is the
latest for that vendor_api. I think this is a little sketch- for eg: if
some other mechanism attempted a watchlist check or KYC call but for
whatever reason didn't save a vres (ie error'd or wasn't run or
something), then the CIP endpoint would not see the latest proper result
we got from that vendor.

This PR just forks that query and makes a version that filters to only
successful, extant vres's and uses that in the CIP endpoint

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->