[FP-6803] Track WorkflowRequests in trigger endpoint (#7055)
## Describe your changes
When a workflow is requested, we will create a WorkflowRequest in the
database. We'll use this to tell when a user still has an outstanding
request, and we'll show this in the dashboard

## Caveats
Haven't quite decided what to do about this: we also support triggering
_only_ a recollect document flow, which is super weird and behaves very
differently from a normal Workflow. I'd like to get rid of this
eventually and just require onboarding onto a playbook.

So, as written, we don't show when there's an outstanding recollect
document flow.

There are a few things we could do:
- Slowly start to phase out the ability to recollect document - just
hide it in the UI from tenants who haven't used it
- _Also_ serialize in `GET /entities` when there's an active Workflow to
collect a document. We'll have to also duplicate that ugly logic in the
filters in the `GET /entities` API. I don't hate this idea for now, but
it's just a bandaid patch. lmk what you think

Will also backfill existing WorkflowRequests with:

```
-- Note, this is omitting adding workflow requests that would be deactivated
WITH new_workflow_requests AS (
    SELECT
        DISTINCT ON(scoped_vault.id)
        scoped_vault.id,
        scoped_vault.fp_id,
        ut1.event->'data'->>'ob_config_id' as ob_config_id,
        ut1.event->'data'->'actor' as actor,
        ut1.timestamp
    FROM user_timeline ut1
    INNER JOIN scoped_vault
        ON ut1.scoped_vault_id = scoped_vault.id
    LEFT JOIN user_timeline ut2
        INNER JOIN onboarding_decision obd
            ON ut2.event->'data'->>'id' = obd.id
        -- Join with timeline events for the same user that show that the user has an onboarding decision
        ON ut2.scoped_vault_id = ut1.scoped_vault_id
        AND ut2.timestamp > ut1.timestamp
        AND ut2.event_kind = 'onboarding_decision'
        AND obd.actor->>'kind' = 'footprint'
    WHERE
        ut1.event_kind = 'workflow_triggered'
        AND ut1.event->'data'->>'ob_config_id' IS NOT NULL
        -- Filter for workflow_triggered events that _do not_ have a later event showing the user has an onboarding decision
        AND ut2.id IS NULL
    ORDER BY 1
)
INSERT INTO workflow_request (scoped_vault_id, ob_configuration_id, timestamp, created_by)
SELECT id, ob_config_id, timestamp, actor FROM new_workflow_requests;

```

## Test Plan
Integration tests