Update backfill_dl script to deactivate speculative data being replaced
We were seeing a case where:
- user started onboarding onto tenant A, added only some data, and then abandoned
- user then fully onboarded onto tenant B. data became portable at tenant B
- tenant A’s view now has both the non-portable data added by tenant A and the portable data added by tenant B.

the fun part: since the user abandoned before finishing onboarding A, it shouldn’t really have access to the data added by tenant B. but current visibility rules would allow this.

in reality, very very few users are affected by this, and to preserve the current behavior, i'm going to backfill these.

this would previously cause an error because prefilling the data from tenant B into tenant A would add two speculative pieces of data with the same DI. this now just deactivates an existing DL when we add new data, just as the normal write path does

Added a new unit test for this case. Before adding the new logic, the unit test would fail with the same error we're seeing in prod