[FP-6733] Use user-specific view to decide what data is portable (#7097)
## Describe your changes
We have this check during portablization to make sure we don't
portablize an ssn4 if there's already an ssn9. Maybe it's overkill now
that we only have one partial CDO :P

Previously, when each tenant-scoped VW would see both speculative data
added by the current tenant _and_ portable data added by other tenants,
we could simply use one VW to compute what was already portable and what
was not.

Soon, the tenant-scoped VW is only going to have data owned by the
current tenant, so it won't know exactly all the data that is already
portable. So, we use the canonical way to compute a VW of only portable
data to now check which data is already portable

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->