[FP-6664] Update view of portable data (#7103)
## Describe your changes
Before tenants had snapshot-isolated views of the world, we could simply
deactivate, for ex, the old address every time we portablized a new
address - every tenant would start seeing the new address as soon as the
old address was deactivated.

In this new world, each tenant has their own view of the world so we
can't mark other tenants' data as deactivated, even if it's replaced by
more portable info. And, in this new world, tenants can decide whenever
they want to delete the data owned by them, and we don't want that to
affect the portable view.

In reality, we want a snapshot-isolated view of the portable data from
the tenants' views of data. We had previously discussed making a
separate "user" vault that stores only portable data for the user - this
is probably something we'll do in the future, but it's another really
involved migration. So in the meantime, going to do this more complex
read-time logic as a patch and catch our breath for a bit :)

This change:
- Ignores the deactivated seqno on DataLifetimes when assembling a
portable view
- Adds some custom logic at read time to properly hide optional DIs that
are replaced by newer data, since we can't use deactivated_at to compute
this

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Unit test

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->