Accept last_activity_at cursors in GET /entities (#7141)
## Describe your changes
To maintain backwards compatibility with the existing GET /entities
requests made by the dashboard, we're going to start by having the
client tell us which kind of cursor it would like to use and we'll
magically adapt.

Notably, I decided to serialize the timestamps as a nanosecond timestamp
instead of isoformat. ISO timestamps display in a really funny way in
the querystring for the dashboard - these nanosecond timestamps show
better

## Caveats
The `GET /users` API uses the ordering_id cursor too... so we can't
deprecate it entirely. Good news is it's marked as a preview API so we
have some justification for making a breaking change if we so choose. I
added some monitoring to see if anyone is using pagination on this API,
my guess is no. In fact, nobody may even be using it

## Test Plan
Integration test, locally in future PR