Fix decryption permissions for optional data (#7150)
## Describe your changes
This is a small bug that was introduced with optional data - if a piece
of data were optional and _not_ in can_access_data, it would _still_,
eroneously, be decryptable. There's a clause in decryption code that
allows any non-playbook-collected data to be decrypted because it then
must have been added via API by the tenant, and optional data was
mistakenly hitting this clause

Probably no impact in prod because can_access_data is already rarely
less than what's collected

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Integration test. This test failed on the old version of code