[1/2] Refactor doc endpoints, functions (#7175)
## Describe your changes
in order to re-use the same APIs for `identity` document_requests and
`proof_of_ssn` document requests, cleaning up some of the code to make
it more easy to follow the shared logic and branch where appropriate.

this is a pure refactor

there are 3 stages of handling a document:
* Create an `IdentityDocument` (`/index` route)
* Upload a side (create a `DocumentUpload`) and associate with an
`IdentityDocument` (`/upload` route)
* upload also processes, but only the appclip uses this part of the
logic, determined via headers. but it's currently not migrated, so we
need to maintain compat for now. i guess we could just wait, but
refactors were needed anyways
* Process the document w/ incode (`/process` route)


in `/index`:
* this is responsible for creating an identity document, which does a
few things:
    * validates the doc type/country is supported
    * sets things like skip_selfie
    * checks if there's a `document_request.kind == 'identity'`
 * **In order to support SSN card:**: we need to 
    * not do the identity specific assertions
    * check doc_request is of type `proof_of_ssn`
    * create the identity doc without any of the identity stuff

in `/upload` 
* we check user consent
* upload to `.latest_image` di
* (appclip process logic)
* generate **some** sandbox fixtures: vres fixtures for alpaca CIP, risk
signals are saved elsewhere
* **In order to support SSN card:**: we need to 
   * save the doc, but not run incode in the temp process flow
   * not save the incode fixtures

in `/process`
* save risk signals fixtures
* run incode
* **In order to support SSN card:**: we need to:
   * not run incode
   * not save fixtures or risk signals

so, given there's a lot of identity document specific code:
- inline and shared between upload/process
- branching from prod<>sandbox
- incode specific

refactoring makes the branching much cleaner and ultimately the code
more clean


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
local :(

i really wanted to write rust tests for these / this stack, but we need
to mock s3 in some fashion which seemed like another detour. I think we
can feel pretty good if we test locally, and have integration tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->