[FP-6614] Generate unique id for each section in docs (#7233)
## Describe your changes
https://onefootprint.slack.com/archives/C03JYNHFEET/p1701969997170609

This fixes a bug where sections with the same name would have the same
anchor tag. Now, we generate the ID for an anchor from the section and
its parent sections.

We use an external package to translate our markdown into JSX, and it
doesn't have a ton of extensibility. So, when we generate the sections
IDs, we actually pass them into this markdown translation package hidden
inside the markdown. We then have our own implementations of H1, H2, and
H3 that pull out these custom IDs from the markdown and render them
properly

## Caveats
This will break any old links to sub-sections

## Test Plan
Locally:


https://github.com/onefootprint/monorepo/assets/21066341/eb6e854c-4a52-4efb-8332-146cbb5ee919