[FP-6873] Allow issuing trigger when workflow is incomplete (#7238)
## Describe your changes
This condition really only existed for the tenant-facing reonboard
endpoint, which does imply that the user has already onboarded. Nobody
is using that API though, so we could even relax that constraint.

Before this, we would show in the dashboard the option to trigger a
workflow request, but it would error if the user were in an incomplete
state. But it's possible the user abandoned and you maybe want to send a
link to ask them to continue onboarding

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Integration test, tested locally