Serialize when there is a workflow request in tenant-facing GET /users/fp_id (#7263)
## Describe your changes
Bloom is going to start opening an FP modal when a user comes back to
their app and has some outstanding information to provide. We have to
first serialize when there's information to collect

## Caveats
This is a public tenant-facing API, so additions should not be made
lightly. It is EXTREMELY difficult to make breaking changes to this API
since we can never be certain who is using them.

For now, I don't feel super confident that WorkflowRequest's inner
interface is something we want to expose publicly. So, i'm keeping it
extremely vague - if there's a workflow requested, we serialize that
there is an `info_requested` object. this object may or may not include
a `note`.

One day, when we feel more confident in our API, maybe we'd serialize
what exactly is requested of the user

~Alternatively, we could gate this serialization on some feature flag
that's only enabled for bloom~
EDIT: i updated to add a preview API feature gate so we can more easily
change this - only bloom will be able to see it :)

## Test Plan
Integration test