Migrate GET /entities to POST /entities/search (#7269)
## Describe your changes
 #7264 but for this GET /entities

Querystrings are logged in many places by default, and these
querystrings may contain PII :(

This migrates `GET /entities?search=` to a POST API that can have an
HTTP body


## Caveats
I initially implemented this as a GET with an HTTP body, but it is not
very conventional at all to have an HTTP GET endpoint that accepts a
body. More info [here](https://www.baeldung.com/http-get-with-body). A
few people disobey this convention, including elastic for their own
"search" API, which is why i was leaning towards it.

But, to my own dismay, i discovered [browsers don't support GETs with an
HTTP body](
https://stackoverflow.com/questions/46404051/send-object-with-axios-get-request)
:( so our dashboard wouldn't be able to use this API. I still have a
slight preference for GET with HTTP body i think. But we need to do this
at least for the dashboard-facing API

## Test Plan
Locally. Will migrate dashboard after this is deployed

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->