Refactor OCR data vaulting to use a common struct (#7258)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Possible solution for how we can solve this for Flexcar:
https://onefootprint.slack.com/archives/C053U235V7G/p1701983884466379

Idea is we want to be able to write a rule that fails with review if
`DocumentOcrNotSuccessful`. Right now, we produce
`DocumentOcrNotSuccessful` purely as a function of the fetch_scores
`id_ocr_confidence.overall_confidence.status`. This is proving
inadequate because this could pass but critical fields like expired_at
and document_number might actually be missing (or low confidence). So
the idea here is to use our own logic to produce this risk signal which
for a given doc type checks all the expected/critical fields and
confirms that they are (1) present and (2) do not have a low confidence
score.

This PR refactors the code we use to write ODK's to the vault so we can
use this same struct for those checks.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
relying on existing tests for this one

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->