Migrate dashboard to use new POST /entities/search (#7268)
## Describe your changes
Querystrings are logged in many places by default, and these
querystrings may contain PII :(

This migrates the dashboard to use the new POST API created in the PR
below this


## Caveats
I initially implemented this as a GET with an HTTP body, but it is not
very conventional at all to have an HTTP GET endpoint that accepts a
body. More info [here](https://www.baeldung.com/http-get-with-body). A
few people disobey this convention, including elastic for their own
"search" API, which is why i was leaning towards it.

But, to my own dismay, i discovered [browsers don't support GETs with an
HTTP body](
https://stackoverflow.com/questions/46404051/send-object-with-axios-get-request)
:( so our dashboard wouldn't be able to use this API. I still have a
slight preference for GET with HTTP body i think. But we need to do this
at least for the dashboard-facing API

## Test Plan
Locally
https://github.com/onefootprint/monorepo/assets/21066341/bc419f4b-0974-4ca0-b12e-02064919cbec