[FP-6650] Ignore 'id already processed' errors (#7309)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

We have a strange case that sometimes happen where our client doesn't
get back a success response from Incode when we call /process/id but
Incode actually did receive and process. Our client then retries and
Incode throws back an error "Id already processed". This used to be a
pretty rare random occurence but suddenly today we are seeing a lot more
cases of it and it is causing a lot of users to get stuck.

This PR is just a quick way for us to recognize and ignore this type of
error and still proceed with the Incode machine as usual. (which is how
we have been manually remedying these cases)

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
tested manually locally 

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->