Write global fingerprints, even for fixture numbers (#7350)
## Describe your changes
We used to not write global fingerprints for fixture numbers to prevent
a situation where:
- tenant A creates a fixture vault on their tenant with sandbox ID A
- tenant B then creates a fixture vault on their tenant with the same
sandbox ID A and one-click logs into the existing vault. This is only
possible because they use the same fixture number, and any tenant can
log into a vault with the fixture number because the code is always
`000000`

But, this was created when sandbox IDs were written manually by the user
instead of randomly generated. Now that they're randomly generated, it's
there's not likely to be a collision.

I didn't love that our fingerprinting logic had this special casing in
many different locations - easier to just create all fingerprints when
possible

## Caveats
If we want to preserve the behavior that tenants still can't log into
each others' fixture vaults, we can add a clause in the identify logic
that _does_ allow logging into other tenants' ONLY for our integration
testing tenants

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->